+++ Goindoor.framework/Versions/A/Headers/OYMBuilding.h
@@ -11,59 +11,48 @@
 
 #import <Foundation/Foundation.h>
 
-#import <Links/Links.h>
+#import "OYMConstant.h"
 
 #import "OYMFloor.h"
+#import "OYMJsonProtocol.h"
+#import "OYMGeometry.h"
+#import "OYMBox.h"
 
 # pragma mark Public constants
 /** String containing the Building type */
 static NSString *const kOYMBuildingType = @"BUILDING";
 
-/** Key to retrieve the building name from the OYMLinksShape */
+/** Key to retrieve the building name */
 static NSString *const kOYMBuildingKeyName = @"name";
-/** Key to retrieve the item type from the OYMLinksShape */
+/** Key to retrieve the floor list */
+static NSString *const kOYMBuildingKeyFloors = @"floors";
+/** Key to retrieve the item type */
 static NSString *const kOYMBuildingKeyType = @"type";
-
+/** Key to retrieve the item id  */
+static NSString *const kOYMBuildingKeyId = @"id";
+static NSString *const kOYMBuildingKeyGeometry = @"geometry";
 
 /**
- *  This class includes all the information regarding the Building Shapes
- * stored in Links. As well, it includes the floors related with the building.
+ *  This class includes all the information regarding the Building stored.
+ * As well, it includes the floors related with the building.
  */
-@interface OYMBuilding : NSObject {
-    @private
-    /** Map containing the user permissions */
-    NSDictionary* acl;
-}
+@interface OYMBuilding : NSObject <OYMJsonProtocol>
 
 #pragma mark Properties
 /** Automatically generated UUID */
 @property (readonly) NSString *uuid;
 /** Building name */
 @property (readonly) NSString *name;
-/** Building {@link Geometry} */
-@property (readonly) Geometry *geometry;
+/** Building Geometry */
+@property (readonly) OYMGeometry *geometry;
 /** String defining Building type */
-@property (readonly) NSString *type;
+@property (readonly) NSString *type DEPRECATED_ATTRIBUTE;
 /** Array including the OYMFloor available in the building */
-@property (readonly) NSMutableDictionary *floors;
-
-
-# pragma mark Constructors
-/**
- *  Building constructor.
- *
- * @param shape OYMLinksShape containing the Building object
- */
-- (instancetype)initWithShape:(OYMLinksShape*)shape;
-
+@property (readonly) NS_MUTABLE_DICTIONARY_OF(NSString*, OYMFloor*) *floors;
+    
 
 # pragma mark Instance methods
-/**
- *  This method creates a OYMLinksShape object from a Building object.
- *
- * @return The equivalent OYMLinksShape object
- */
-- (OYMLinksShape*)toShape;
+
 /**
  *  This method adds a OYMFloor in the Building's Floor list.
  *
@@ -79,9 +68,9 @@ static NSString *const kOYMBuildingKeyType = @"type";
 /**
  *  Getter for the Floor list.
  *
- * @return An <tt>ArrayList&lt;Floor&gt;</tt> ordered per floor number (smaller first)
+ * @return An Array of OYMFloor, ordered by floor number (smaller first)
  */
-- (NSArray*) getFloorsList;
+- (NS_ARRAY_OF(OYMFloor*)*) getFloorsList;
 
 
 @end