+++ Goindoor.framework/Versions/A/Headers/OYMFloor.h
@@ -12,26 +12,29 @@
 
 #import <Foundation/Foundation.h>
 
-#import <Links/links.h>
+#import "OYMConstant.h"
 
 #import "OYMTileOverlay.h"
+#import "OYMJsonProtocol.h"
 
 /** String containing the Floor type */
 static NSString *const kOYMFloorType = @"FLOOR";
 
-/** Key to retrieve the building id from the OYMLinksItem */
+/** Key to retrieve the building id */
 static NSString *const kOYMFloorKeyBuilding = @"building";
-/** Key to retrieve the floor number from the OYMLinksItem */
-static NSString *const kOYMFloorKeyFloor = @"floornumber";
-/** Key to retrieve the floor type from the OYMLinksItem */
+/** Key to retrieve the floor number */
+static NSString *const kOYMFloorKeyFloor = @"floorNumber";
+/** Key to retrieve the floor type */
 static NSString *const kOYMFloorKeyType = @"type";
+static NSString *const kOYMFloorKeyId = @"id";
+
 
 /**
- *  This class includes all the information regarding the Floor Items stored
- * in Links. As well, it includes the GMSURLTileLayer to be used to
- * overlap the Goindoor maps in the Google Map.
+ *  This class includes all the information regarding the Floor Items stored.
+ * As well, it includes the UrlTileProvider to be used to
+ * overlap the indoor maps in the Map
  */
-@interface OYMFloor : NSObject
+@interface OYMFloor : NSObject <OYMJsonProtocol>
 
 #pragma mark Properties
 /** Automatically generated UUID */
@@ -41,19 +44,9 @@ static NSString *const kOYMFloorKeyType = @"type";
 /** Floor number */
 @property (readonly) NSNumber *floor;
 /** String defining Floor type */
-@property (readonly) NSString *type;
+@property (readonly) NSString *type DEPRECATED_ATTRIBUTE;
 /** Tile provider for Goindoor tiles */
 @property (readonly) OYMTileOverlay *tileProvider;
-
-
-# pragma mark Constructors
-/**
- *  Floor constructor.
- *
- * @param item OYMLinksItem containing the Floor object
- */
-- (instancetype)initWithItem:(OYMLinksItem*)item;
-
 
 @end
 #endif