+++ Goindoor.framework/Versions/A/Headers/goindoor.h
@@ -8,6 +8,9 @@
 
 #import <Foundation/Foundation.h>
 
+#import <Goindoor/OYMReachability.h>
+#import <Goindoor/OYMConstant.h>
+
 #import <Goindoor/OYMBuilding.h>
 #import <Goindoor/OYMFloor.h>
 #import <Goindoor/OYMTileOverlay.h>
@@ -22,9 +25,20 @@
 #import <Goindoor/OYMLogger.h>
 #import <Goindoor/OYMAsset.h>
 #import <Goindoor/OYMDataServer.h>
-#import <Goindoor/OYMDataLinks.h>
 #import <Goindoor/OYMDataHandler.h>
 #import <Goindoor/OYMGoindoor.h>
+
+#import <Goindoor/OYMWebservice.h>
+#import <Goindoor/OYMDataWebservice.h>
+#import <Goindoor/OYMUtils.h>
+#import <Goindoor/OYMProxiBeacon.h>
+
+#import <Goindoor/OYMGeometry.h>
+#import <Goindoor/OYMBox.h>
+#import <Goindoor/OYMGeometryPoint.h>
+#import <Goindoor/OYMCircle.h>
+#import <Goindoor/OYMPolygon.h>
+#import <Goindoor/OYMPolyline.h>
 
 #import <Goindoor/OYMLocationCore.h>
 #import <Goindoor/OYMLocationDelegate.h>