+++ Sample app/Frameworks/Goindoor.framework/Versions/A/Headers/OYMBeacon.h
@@ -13,11 +13,12 @@
 #import <CoreLocation/CoreLocation.h>
 #import <CoreBluetooth/CoreBluetooth.h>
 
-#import <Links/links.h>
+#import "OYMConstant.h"
 
 #import "NSData+Hex.h"
 #import "ESSEddystone.h"
 
+#import "OYMJsonProtocol.h"
 
 # pragma mark Public constants
 /** String containing the Beacon type */
@@ -44,40 +45,51 @@ static NSString *const kOYMBeaconEddystoneHeader2 = @"16AAFE";
 /** String containing the Eddystone-UID Frame Type */
 static NSString *const kOYMBeaconEddystoneFrameUID = @"00";
 
-/** Key to retrieve the Beacon major from the {@link Item} */
+/** Key to retrieve the Beacon uuid */
+static NSString *const kOYMBeaconKeyId = @"id";
+/** Key to retrieve the Beacon major */
 static NSString *const kOYMBeaconKeyMajor = @"major";
-/** Key to retrieve the Beacon minor from the {@link Item} */
+/** Key to retrieve the Beacon minor */
 static NSString *const kOYMBeaconKeyMinor = @"minor";
-/** Key to retrieve the Beacon transmission power from the {@link Item} */
+/** Key to retrieve the Beacon longitude */
+static NSString *const kOYMBeaconKeyLongitude = @"longitude";
+/** Key to retrieve the Beacon latitude */
+static NSString *const kOYMBeaconKeyLatitude = @"latitude";
+/** Key to retrieve the Beacon transmission power */
 static NSString *const kOYMBeaconKeyTx = @"tx";
-/** Key to retrieve the Beacon proximity UUID from the {@link Item} */
-static NSString *const kOYMBeaconKeyUuid = @"proximityUUID";
-/** Key to retrieve the Beacon {@link Building} ID from the {@link Item} */
+/** Key to retrieve the Beacon proximity UUID */
+static NSString *const kOYMBeaconKeyUuid = @"uuid";
+/** Key to retrieve the Beacon Building ID */
 static NSString *const kOYMBeaconKeyBuilding = @"building";
-/** Key to retrieve the Beacon {@link Floor} ID from the {@link Item} */
+/** Key to retrieve the Beacon Floor ID */
 static NSString *const kOYMBeaconKeyFloor = @"floor";
-/** Key to retrieve the Beacon {@link Floor} number from the {@link Item} */
-static NSString *const kOYMBeaconKeyFloorNumber = @"floornumber";
-/** Key to retrieve the Beacon type from the {@link Item} */
+/** Key to retrieve the Beacon Floor number */
+static NSString *const kOYMBeaconKeyFloorNumber = @"floorNumber";
+/** Key to retrieve the Beacon type */
 static NSString *const kOYMBeaconKeyType = @"type";
-/** Key to retrieve the Beacon status from the {@link Item} */
+/** Key to retrieve the Beacon status */
 static NSString *const kOYMBeaconKeyStatus = @"status";
-/** Key to retrieve the Beacon warning counter from the {@link Item} */
+/** Key to retrieve the Beacon warning counter */
 static NSString *const kOYMBeaconKeyWarningCounter = @"warningCounter";
+/** Key to retrieve the Beacon ScanDate */
+static NSString *const kOYMBeaconKeyScanDate = @"scanDate";
 
 /**
- *  This class includes all the information regarding the Beacon Items stored
- * in Links.
+ *  This class includes all the information regarding the Beacon Items stored.
  */
-@interface OYMBeacon : NSObject
+@interface OYMBeacon : NSObject <OYMJsonProtocol>
 
 #pragma mark Properties
-/** OYMLinksItem ID */
+/** Automatically generated UUID */
 @property NSString *uuid;
 /** WGS84 Longitude */
-@property NSNumber *x;
+@property (nonatomic) NSNumber *x DEPRECATED_MSG_ATTRIBUTE("Use longitude instead.");
 /** WGS84 Latitude */
-@property NSNumber *y;
+@property (nonatomic) NSNumber *y  DEPRECATED_MSG_ATTRIBUTE("Use latitude instead.");
+/** WGS84 Longitude */
+@property NSNumber *longitude;
+/** WGS84 Latitude */
+@property NSNumber *latitude;
 /** Beacon major */
 @property NSNumber *major;
 /** Beacon minor */
@@ -93,7 +105,7 @@ static NSString *const kOYMBeaconKeyWarningCounter = @"warningCounter";
 /** Floor number */
 @property NSNumber *floorNumber;
 /** String defining the Beacon type */
-@property NSString *type;
+@property NSString *type DEPRECATED_ATTRIBUTE;
 /** Scan Date */
 @property NSString *scanDate;
 /** Variable defining the Beacon status, if &lt;1 can be used */
@@ -102,12 +114,7 @@ static NSString *const kOYMBeaconKeyWarningCounter = @"warningCounter";
 @property int warningCounter;
 
 #pragma mark Constructors
-/**
- *  Beacon constructor.
- *
- * @param item OYMLinksItem containing the Beacon object
- */
-- (instancetype)initWithItem:(OYMLinksItem*)item;
+
 /**
  *  Eddystone Beacon partial constructor. Should not be used.
  *
@@ -133,12 +140,6 @@ static NSString *const kOYMBeaconKeyWarningCounter = @"warningCounter";
  * @return A unique Beacon identifier
  */
 - (NSString*)getIdentifier;
-/**
- *  This method creates an OYMLinksItem object from an Beacon object.
- *
- * @return The equivalent OYMLinksItem object
- */
-- (OYMLinksItem*)toItem;
 
 @end
 #endif