+++ Sample app/Navigation/AppDelegate.h
@@ -9,7 +9,6 @@
 #import <UIKit/UIKit.h>
 
 #import "Delegate.h"
-#import "Reachability.h"
 
 
 @interface AppDelegate : UIResponder <UIApplicationDelegate>