+++ Sample app/Navigation/Views/OYMTransparentTextField.m
@@ -77,4 +77,8 @@
     return returnValue;
 }
 
+- (BOOL)canPerformAction:(SEL)action withSender:(id)sender {
+    return NO;
+}
+
 @end